V2 convert helpers to extension methods
