+++ Model/RegistrationResult.cs
@@ -1,4 +1,7 @@
-﻿namespace HackerFerret.ScormHelper.Model
+﻿using RusticiSoftware.HostedEngine.Client;
+using System.Xml;
+
+namespace HackerFerret.ScormHelper.Model
 {
     public class RegistrationResult
     {
@@ -6,9 +9,6 @@
         {
         }
 
-        public RegistrationResult(string xmlVal)
-        {
-        }
 
         public string Title { get; set; }
         public string Complete { get; set; }
@@ -16,8 +16,8 @@
         public string Score { get; set; }
         public string Attempts { get; set; }
         public string ViewTime { get; set; }
-
     }
+
 }
