+++ ScormHelper.csproj
@@ -9,7 +9,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>HackerFerret.ScormHelper</RootNamespace>
     <AssemblyName>ScormHelper</AssemblyName>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <SccProjectName>SAK</SccProjectName>
     <SccLocalPath>SAK</SccLocalPath>
@@ -37,10 +37,6 @@
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="RusticiSoftware.HostedEngine.Client, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>packages\RusticiSoftware.HostedEngine.Client.1.1.0.0\lib\net461\RusticiSoftware.HostedEngine.Client.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -51,23 +47,28 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Api\CourseApi.cs" />
-    <Compile Include="Api\ReportApi.cs" />
+    <Compile Include="Extensions\ReportingServiceExtensions.cs" />
+    <Compile Include="Extensions\CourseServiceExtensions.cs" />
+    <Compile Include="Extensions\RegistrationServiceExtensions.cs" />
     <Compile Include="Extensions\CourseExtensions.cs" />
     <Compile Include="Extensions\RegistrationExtensions.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Helpers\Converter.cs" />
-    <Compile Include="Helpers\RegistrationResultParser.cs" />
     <Compile Include="Model\RegistrationResult.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Api\RegistrationApi.cs" />
-    <Compile Include="Api\Common.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="package.nuspec" />
     <None Include="packages.config" />
     <None Include="Readme.md" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SCORMCloud_NetLibrary\RusticiSoftware.HostedEngine.Client.csproj">
+      <Project>{11af12ee-c99e-4a73-8be2-1fc016867bc0}</Project>
+      <Name>RusticiSoftware.HostedEngine.Client</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
